preprocessing/requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
numpy
scikit-learn